removed Image.ios
