--- packages/palette/src/elements/Image/Image.ios.tsx
@@ -1,9 +1,0 @@
-import { BaseImage, BaseResponsiveImage } from "./Image.shared"
-
-/** Image */
-export const Image = BaseImage
-Image.displayName = "Image"
-
-/** ResponsiveImage */
-export const ResponsiveImage = BaseResponsiveImage
-ResponsiveImage.displayName = "ResponsiveImage"
